+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This release brings a major refactor of the forms and fields api. Its aim is to 
 - Removed: extract dynamic fields based on field definitions.
 - Removed: chiefRegister()->model() is removed. Use chiefRegister()->resource().
 - Changed: Each viewable model needs to have a `viewKey()` method to determine the frontend view path.
+- Changed: Nav tags are no longer added via the register method, chiefRegister()->resource(). Instead on the page resource you should set the `protected function getNavTags(): array` method to provide the nav tags.
 
 - Changed: Restructured Fields files and classes under a dedicated `src/Forms` directory. Rename all your field namespaces from `Thinktomorrow\Chief\ManagedModels\Fields\Types\<Field>` to `Thinktomorrow\Chief\Forms\Fields\<Field>`.
 - Removed: field method `customSaveMethod`. Replace any occurrences with the save() method. See below.